Frambuffer: API changes for ease of use
Moved frambuffer into its down file and changes how it is created and
use. The shader make now create mono::local<> instead of mono::ref<>,
this is because it can be converted to mono::ref<> easily but not the
other way around.

Remove mipmap and param methods in texture. The api is clunky to use and
it is better to just use the glTexParamteri directly for now.