Event types and registrations
window_icon, window_maximize and content_scale events are added.
window class interface for settting the position. The event
registrations are semi sorted to how the event_type enum is defined.

Fixes where the `auto it` might actually be a copy by using
`auto const&` instead to have the it->second be a const reference.

Remove draw interface from graphics_context, this might be added in
later but for now it is not needed.